taxes-backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
passlib[bcrypt]==1.7.4
python-multipart
alembic
python-jose[cryptography]
bcrypt==4.0.1
pydantic[email]
